Fix type for perf event config getter
REVIEW: 11007163 (cherry-pick from ce7f5c0c77c0bf543596a8962fa5e97f7cc00036)
commit_hash:efe97f83fcdc5da6c83dc7e3e8f9206281a0cf6c

+++ perforator/agent/collector/progs/unwinder/unwinder.bpf.c
@@ -276,7 +276,7 @@ static ALWAYS_INLINE u32 get_perf_event_type(struct bpf_perf_event_data* ctx) {
     return BPF_CORE_READ(kctx, event, attr.type);
 }
 
-static ALWAYS_INLINE u32 get_perf_event_config(struct bpf_perf_event_data* ctx) {
+static ALWAYS_INLINE u64 get_perf_event_config(struct bpf_perf_event_data* ctx) {
     struct bpf_perf_event_data_kern* kctx = (void*)ctx;
     return BPF_CORE_READ(kctx, event, attr.config);
 }
